+++ History.md
@@ -1,3 +1,13 @@
+# idb-batch
+
+## ?
+
+* Breaking change (minor): Use `sync-promise`;
+* Added `getStoreNames`
+* Feature: Support `transactionalBatch` (incomplete);
+* Feature: Allow preexisting transaction to be supplied to `batch()` (or `transactionalBatch()`);
+* Feature: Support `move`, `copy` operations;
+
 ## 1.0.0 / 2015-12-06
 
 * initial release :sparkles: